[MID-36] move filters to Specialist folder
